% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{dfs_from_gam}
\alias{dfs_from_gam}
\title{extract degrees of freedom from a gam object}
\usage{
dfs_from_gam(fit)
}
\arguments{
\item{fit}{an object returned by gam()}
}
\description{
extract degrees of freedom from a gam object
}
