% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{extract_formula}
\alias{extract_formula}
\title{extract a formula from a gam object 
extract estimated degrees of freedom from a gam object and return a formula
  that can serve as an input for a gamlss estimation}
\usage{
extract_formula(fit)
}
\description{
extract a formula from a gam object 
extract estimated degrees of freedom from a gam object and return a formula
  that can serve as an input for a gamlss estimation
}
