% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_gamlss_bootstrap.R
\name{ImpGamlssBootstrap}
\alias{ImpGamlssBootstrap}
\title{GAMLSS bootstrap method}
\usage{
ImpGamlssBootstrap(incomplete.data, fit, R, ...)
}
\arguments{
\item{incomplete.data}{Data frame with missings on one variable.}

\item{fit}{Random sample generator method.}

\item{R}{Boolean matrix with the response indicator.}

\item{...}{extra arguments for the control of the gamlss fitting
function}
}
\value{
Returns a imputation sample generator.
}
\description{
Creates a random generation function for the missing
  values with bootstrap sample from the fitted GAMLSS model for the
  completely observed data.
}

