% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_gamlss_fit.R
\name{ImpGamlssFit}
\alias{ImpGamlssFit}
\title{GAMLSS imputation fit}
\usage{
ImpGamlssFit(data, new.data, family, n.ind.par, gd.tol, lin.terms, forceNormal,
  c.crit, n.cyc, trace, ...)
}
\arguments{
\item{data}{Completely observed data frame to be used to fit a
gamlss model estimate.}

\item{new.data}{Data frame used to predict the parameter values for
some given right side x-values on the gamlss model.}

\item{family}{Family to be used for the response variable on the
GAMLSS estimation.}

\item{n.ind.par}{Number of individual parameters to be
fitted. Currently it only allows one or two because of stability
issues for more parameters.}

\item{gd.tol}{Global tolerance of the gamlss fitting function.}
}
\value{
Returns a method to generate random samples for the fitted
  gamlss model using "new.data" as covariates.
}
\description{
This function takes uses a data set to fit a gamlss model and
another to predict the expected parameters values. It returns a
function that will generate a vector of ramdom observations for the
predicted parameters. The amount of random obervations is the
number of units on the dataset used to get such predictions.
}

