% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_creator.R
\name{ModelCreator}
\alias{ModelCreator}
\title{Model creator}
\usage{
ModelCreator(data, gam.model, lin.terms = NULL)
}
\arguments{
\item{data}{Data frame that will provide the named variables.}

\item{gam.model}{List of mode parameter, containing the "type" with
c("linear", "cs", "pb") as available choices and "par", an
optional list parameter if the model is not linear.}

\item{lin.terms}{Specify which predictors should be included
linearly. For example, binary variables can be added directly as
an additive term instead of defining a spline.}
}
\value{
Returns a formula object.
}
\description{
This is a helper function to be used within the gamlss
  fitting procedure. It creates automatically a formula object for
  the variables named a given data frame. The dependent variable is
  the one in the first column and the rest are treated as
  independent.
}
