% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{sum_edf}
\alias{sum_edf}
\title{sum all estimated degrees of freedom that belong to same coefficient}
\usage{
sum_edf(coef, edf_list)
}
\arguments{
\item{coef}{a coefficient name from a gam}

\item{edf_list}{a list with all estimated degrees of freedom}
}
\description{
sum all estimated degrees of freedom that belong to same coefficient
}
