% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{term_to_str_formula}
\alias{term_to_str_formula}
\title{collapse terms into a formula}
\usage{
term_to_str_formula(elements, response)
}
\description{
collapse terms into a formula
}
\examples{
elements <- dfs_to_term(dfs_from_gam(fit))
term_to_str_formula(elements, "pollen")
}
